Fix Git Terminal Lifecycle And Custom Command Support
Keep the Git plugin’s terminal title stable, allow it to open a terminal with an explicit shell command, and close any Git-labeled terminal when the workspace is shut down.

Update the builtin terminal pane so callers can override the terminal command even when no terminal types are configured, while preserving the existing selection-based behavior for normal launches. Also add a helper to close a terminal tab by title and use it from the Git plugin cleanup path.

Tighten an SFTP browser icon lookup to avoid signed/unsigned size comparison issues.

* Git plugin
* Builtin terminal pane
* SFTP browser

+++ Plugin/SFTPBrowserDlg.cpp
@@ -217,8 +217,8 @@ void SFTPBrowserDlg::DoDisplayEntriesForPath(const wxString& path)
                     imgid = loader->GetMimeImageId(fn.GetFullName());
                 }
             }
-            cols.push_back(
-                ::MakeIconText(attr->GetName(), bitmaps->size() > imgid ? bitmaps->at(imgid) : wxNullBitmap));
+            cols.push_back(::MakeIconText(
+                attr->GetName(), static_cast<int>(bitmaps->size()) > imgid ? bitmaps->at(imgid) : wxNullBitmap));
             cols.push_back(attr->GetTypeAsString());
             cols.push_back(wxString() << attr->GetSize());
